+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [Ruby,JavaScript,Go] Add `bigdecimal`, `biginteger` parameter types ([#42](https://github.com/cucumber/cucumber-expressions/pull/42))
 - [.NET] Implementation of Cucumber Expressions by porting the Java parser
 ([#1743](https://github.com/cucumber/cucumber-expressions/pull/45))
+- [Python] Added Python Cucumber Expressions
+([#65](https://github.com/cucumber/cucumber-expressions/pull/65))
 
 ### Changed
 - [Go] Parameters of type `{float}` are now parsed as `float32` (previously it was `float64`).